Move connectivity_t and related types out of mesh_types.h
